Refactor assignment models: move to basemodels.py and update structure
