+++ basemodels.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from typing import List, Optional
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 class TokenRequest(BaseModel):
     token: str
@@ -70,6 +70,7 @@ class Message(BaseModel):
     class Config:
         from_attributes = True
 
+# TODO: fix like in AssignmentResponse
 class MessageResponse(BaseModel):
     message: str
     room_id: int
@@ -81,5 +82,36 @@ class MessageResponse(BaseModel):
     _id: str
 
 class MessagesRetrieve(BaseModel):
+    room_id: int
+    user_token: str
+
+
+
+class Assignment(BaseModel):
+    message: str
+    user_token: str
+    room_id: int
+    is_private: bool
+    reply_to: Optional[str] = None
+    attachments: Optional[List[str]] = None  # List of file URLs or filenames
+
+    class Config:
+        from_attributes = True
+
+class AssignmentResponse(BaseModel):
+    id: str = Field(..., alias="_id")
+    message: str
+    room_id: int
+    is_private: bool
+    reply_to: Optional[str] = None
+    user_id: int
+    timestamp: datetime
+    grade: Optional[int] = None
+    attachments: Optional[List[str]] = None
+
+    class Config:
+        allow_population_by_field_name = True
+
+class AssignmentsRetrieve(BaseModel):
     room_id: int
     user_token: str
